+++ NEWS.md
@@ -1,3 +1,8 @@
+# rhino (development version)
+
+1. Add Rstudio addins for lint, build and test Sass, R and Javascript. Updated new module addin.
+
+
 # rhino 1.4.0
 
 1. New `pkg_install()` and `pkg_remove()` functions to simplify dependency management in Rhino.